sdl3: continue API updates and iOS window handling
Add iOS build defines/warning suppressions in CMake, update SDL3 API
usage across game, DB, and input code, and add iOS orientation/window
handling with svga debug logging.

Why: keep the SDL3 migration compiling and test iOS behavior.

+++ src/fps_limiter.cc
@@ -13,7 +13,7 @@ FpsLimiter::FpsLimiter(unsigned int fps)
 
 void FpsLimiter::mark()
 {
-    _ticks = SDL_GetTicks();
+    _ticks = static_cast<unsigned int>(SDL_GetTicks());
 }
 
 void FpsLimiter::throttle() const
@@ -22,8 +22,8 @@ void FpsLimiter::throttle() const
         return;
     }
 
-    if (1000 / _fps > SDL_GetTicks() - _ticks) {
-        SDL_Delay(1000 / _fps - (SDL_GetTicks() - _ticks));
+    if (1000 / _fps > static_cast<unsigned int>(SDL_GetTicks()) - _ticks) {
+        SDL_Delay(1000 / _fps - (static_cast<unsigned int>(SDL_GetTicks()) - _ticks));
     }
 }
